+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - bump blockfrost-utils to 2.6.2
+- bump node to 18 in shell.nix
 
 ## [1.6.0] - 2023-06-28